Separate pixel classifier advanced options
